fixed several golint warnings
