+++ kyt-alm-server/go.mod
@@ -5,7 +5,7 @@ go 1.16
 require (
 	github.com/amenzhinsky/iothub v0.6.2
 	github.com/appleboy/gin-jwt/v2 v2.6.4
-	github.com/ci4rail/kyt/kyt-server-common/iothub_wrapper v0.0.0-00010101000000-000000000000
+	github.com/ci4rail/kyt/kyt-server-common/iothubwrapper v0.0.0-00010101000000-000000000000
 	github.com/ci4rail/kyt/kyt-server-common/token v0.0.0-00010101000000-000000000000
 	github.com/ci4rail/kyt/kyt-server-common/version v0.0.0-00010101000000-000000000000
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible
@@ -18,6 +18,6 @@ require (
 
 replace github.com/ci4rail/kyt/kyt-server-common/token => ../kyt-server-common/token
 
-replace github.com/ci4rail/kyt/kyt-server-common/iothub_wrapper => ../kyt-server-common/iothub_wrapper
+replace github.com/ci4rail/kyt/kyt-server-common/iothubwrapper => ../kyt-server-common/iothubwrapper
 
 replace github.com/ci4rail/kyt/kyt-server-common/version => ../kyt-server-common/version
