+++ kyt-server-common/iothubwrapper/iothub_service.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package iothub_wrapper
+package iothubwrapper
 
 import (
 	"context"
